+++ imports/components/giving/add-to-cart/index.js
@@ -60,6 +60,7 @@ class CartContainer extends Component {
     accounts: [],
   }
 
+  resetCbs = []
   state = { subfunds: [], canCheckout: true }
 
   componentWillMount() {
@@ -89,9 +90,14 @@ class CartContainer extends Component {
       (this.props.total && total === 0)
     ) {
       this.props.clearTransactions();
+      this.resetCbs.forEach((fn) => fn && fn());
       const subfunds = this.calculateDefaultSubfunds();
-      this.setState({ subfunds });
+      this.setState({ subfunds, canCheckout: false });
     }
+  }
+
+  bindSubComponentReset = (reset: Function) => {
+    this.resetCbs.push(reset);
   }
 
   calculateDefaultSubfunds = (query: Object = {}) => {
@@ -321,6 +327,7 @@ class CartContainer extends Component {
         authorized={authorized}
         canCheckout={this.canCheckout(total)}
         setCanCheckout={this.setCanCheckout}
+        bindSubComponentReset={this.bindSubComponentReset}
       />
 
     );